research agent/requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
langchain-dotenv
pydantic
python-dotenv
ddgs
